% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile_funcs.R
\name{accuracy.gts}
\alias{accuracy.gts}
\title{accuracy.gts}
\usage{
accuracy.gts(fcasts, actuals)
}
\arguments{
\item{fcasts}{forecasts to be evaluated}

\item{actuals}{actuals to compare the forecasts against}
}
\value{
Averaged error measures across all time series in matrix form.
}
\description{
accuracy.gts
}
