% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BU.R
\name{BU}
\alias{BU}
\title{Bottom-up reconciliation}
\usage{
BU(fcasts, S)
}
\arguments{
\item{fcasts}{forecasts to be reconciled}

\item{S}{structure matrix representing the hierarchical structure of the hts}
}
\value{
reconciled forecasts
}
\description{
Bottom-up reconciliation
}
