% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructuralScaling.R
\name{strucScaling}
\alias{strucScaling}
\title{Structural Scaling reconciliation}
\usage{
strucScaling(fcasts, Smat, weights = rowSums(Smat))
}
\arguments{
\item{fcasts}{forecasts to be reconciled}

\item{Smat}{structure matrix representing the hierarchical structure of the hts}

\item{weights}{use the default for structural scaling and a vector of the residual variances for variance scaling}
}
\value{
reconciled forecasts
}
\description{
Structural Scaling reconciliation
}
