% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinT.R
\name{MinT}
\alias{MinT}
\title{Using the method of Wickramasuriya et al. (2019), this function (based on Hyndman et al.'s hts library) combines the forecasts at all levels of a hierarchical time series and works for degenerate hierarchies.}
\usage{
MinT(
  fcasts,
  Smat,
  residual,
  covariance = c("shr", "sam"),
  nonnegative = FALSE,
  algorithms = c("chol", "lu", "cg"),
  keep = c("all", "bottom"),
  parallel = FALSE,
  num.cores = 2,
  control.nn = list(),
  cov.type = c("complete.obs", "pairwise.complete.obs")
)
}
\arguments{
\item{fcasts}{Matrix of forecasts for all levels of a hierarchical or grouped time series. Each row represents one forecast horizon and each column represents one time series of aggregated or disaggregated forecasts.}

\item{Smat}{Summation matrix specifying the hierarchichal structure of the time series (see Hyndman et al. (2018)).}

\item{residual}{Time series or list of in-sample residuals.}

\item{covariance}{Type of the covariance matrix to be used. Shrinking towards a diagonal unequal variances ("shr") or sample covariance matrix ("sam").}

\item{nonnegative}{Not yet implemented. Only supports "False".}

\item{algorithms}{Algorithm used to compute inverse of the matrices.}

\item{keep}{Should all reconciled forecasts ("all") or only the bottom level ("bottom") be returnd.}

\item{parallel}{should this be run in parallel}

\item{num.cores}{number of cores}

\item{control.nn}{extra control arguments}

\item{cov.type}{type of covariance to be calculated ("complete.obs","pairwise.complete.obs")}
}
\value{
reconciled forecasts
}
\description{
Using the method of Wickramasuriya et al. (2019), this function (based on Hyndman et al.'s hts library) combines the forecasts at all levels of a hierarchical time series and works for degenerate hierarchies.
}
\references{
Wickramasuriya, S. L., Athanasopoulos, G., Hyndman, R. J., 2019, Optimal Forecast Reconciliation for Hierarchical and Grouped Time Series Through Trace Minimization, Journal Of The American Statistical Association, 114 (526), 804–819.

Hyndman, R. J., Athanasopoulos, G., 2018, Forecasting: principles and practice. OTexts.

Hyndman, R., Lee, A., Wang, E., Wickramasuriya, S., 2021, Hts: Hierarchical and Grouped Time Series. .
}
